+++ app/src/main/java/ru/vdv/myapp/myreadersdiary/glide/GlideImageLoader.kt
@@ -17,7 +17,6 @@ class GlideImageLoader : IImageLoader<ImageView> {
             .centerCrop()
             .into(container)
     }
-
     fun loadInfo(url: String, container: ImageView) {
         Glide.with(container.context)
             .load(url)
@@ -39,6 +38,7 @@ class GlideImageLoader : IImageLoader<ImageView> {
             //.circleCrop()
             .into(container)
     }
+
 
 
 }
