--- app/src/main/java/ru/vdv/myapp/myreadersdiary/interactors/ListEventInteractor.kt
@@ -1,4 +1,0 @@
-package ru.vdv.myapp.myreadersdiary.interactors
-
-class ListEventInteractor {
-}
