+++ app/src/main/java/ru/vdv/myapp/myreadersdiary/model/api/ApiService.kt
@@ -1,16 +1,72 @@
 package ru.vdv.myapp.myreadersdiary.model.api
 
-/**
-## [Dad Approves API Docs](https://dadapproves.ru/docs/overview-resources-in-the-rest-api.php)
-### Сводный (обобщающий) интерфейс API Service
-подключенные интерфейсы компонент:
-- [DApiUsers] - интерфейс компоненты [API USERS (книги)](https://dadapproves.ru/docs/overview-resources-in-the-rest-api.php)
-- [DApiBooks]  - интерфейс компоненты [API BOOKS (Книги)](https://dadapproves.ru/docs/overview-resources-in-the-rest-api.php)
-- [DApiEvents] - интерфейс компоненты [API EVENTS (События)](https://dadapproves.ru/docs/overview-resources-in-the-rest-api.php)
- */
-interface ApiService : DApiBooks, DApiEvents, DApiUsers {
-    // в теле интерфейса ммогут быть размещены методы общего характера, тестовые и не относящиеся
-    // явно к конкретным компонентам сервиса. Если метод стандартный для компонента, имеет
-    // ссылку н подробное описание и оформлен в полном соответствии с документацией
-    // желательно его размещать в соответствующем файле интерфейса
+import retrofit2.Call
+import retrofit2.http.*
+import ru.vdv.myapp.myreadersdiary.domain.Book
+import ru.vdv.myapp.myreadersdiary.domain.User
+
+interface ApiService {
+    /**
+     * Режим: API/BOOKS      - при таком запросе запрашивается список книг пользователя
+     * @param user ......... - Идентификатор пользователя, список книг которого запрашивается
+     * @param key .......... - базовый ключ пользователя API key
+     * @param page ......... - номер запрашиваемой страницы ( >=1 )
+     * @return возвращает список книг пользователя...
+     */
+
+    //для целей тестирования реализованы 8 основных методов (REST протоколов)
+    @GET("books")
+    fun getListOfBooks(
+        @Header("apiKey") key: String,
+        @Header("user") user: String,
+        @Header("page") page: Int,
+    ): Call<List<Book>>
+
+    //для целей тестирования реализованы 8 основных методов (REST протоколов)
+    @GET("user")
+    fun getUserInfo(
+        @Header("apiKey") key: String,
+        @Header("daUserLogin") user: String,
+    ): Call<User>
+
+    @POST("books")
+    fun postBook(
+        @Header("dauser") user: String,
+        @Body book: Book
+    ): Call<Any>
+
+    @PATCH("books")
+    fun patchBook(
+        @Header("dauser") user: String,
+        @Body book: Book
+    ): Call<Any>
+
+    @DELETE("books")
+    fun deleteBook(
+        @Header("dauser") user: String,
+    ): Call<Any>
+
+    @PUT("books")
+    fun putBook(
+        @Header("dauser") user: String,
+        @Body book: Book
+    ): Call<Any>
+
+    @HEAD("books")
+    fun headBook(
+        @Header("dauser") user: String,
+        @Body book: Book
+    ): Call<Any>
+
+    @OPTIONS("books")
+    fun optionsBook(
+        @Header("dauser") user: String,
+        @Body book: Book
+    ): Call<Any>
+
+    @HTTP(method = "get", path = "books", hasBody = false)
+    fun httpBook(
+        @Header("dauser") user: String,
+        @Body book: Book
+    ): Call<Any>
 }
