+++ app/src/main/java/ru/vdv/myapp/myreadersdiary/ui/book/BookDetailsFragment.kt
@@ -37,9 +37,7 @@ class BookDetailsFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         book = arguments?.getParcelable("ARG_BOOK")!!
         binding.textViewBookDetailsTitle.text = book.title
-        "${book.producerName} ${book.producerPatronymic} ${book.producerSurname}".also {
-            binding.textViewBookDetailsAuthorsOfTheBook.text = it
-        }
+        "${book.producerName} ${book.producerPatronymic} ${book.producerSurname}".also { binding.textViewBookDetailsAuthorsOfTheBook.text = it }
         imageLoader.loadBookCover(
             "https://dadapproves.ru/usercontent/book/covers/${book.bookCover}",
             binding.imageViewBookDetailsCover
@@ -48,4 +46,5 @@ class BookDetailsFragment : Fragment() {
         viewModel = ViewModelProvider(this).get(BookDetailsViewModel::class.java)
         // TODO: Use the ViewModel
     }
+
 }
