+++ app/src/main/java/ru/vdv/myapp/myreadersdiary/ui/book/CreateNewBookFragment.kt
@@ -28,4 +28,5 @@ class CreateNewBookFragment : Fragment() {
         viewModel = ViewModelProvider(this).get(CreateNewBookViewModel::class.java)
         // TODO: Use the ViewModel
     }
+
 }
