+++ app/src/main/java/ru/vdv/myapp/myreadersdiary/ui/main/MainViewModel.kt
@@ -3,9 +3,7 @@ package ru.vdv.myapp.myreadersdiary.ui.main
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import ru.vdv.myapp.myreadersdiary.domain.Book
 import ru.vdv.myapp.myreadersdiary.domain.CallBack
-import ru.vdv.myapp.myreadersdiary.domain.Event
 import ru.vdv.myapp.myreadersdiary.domain.User
 import ru.vdv.myapp.myreadersdiary.model.repository.RepositoryImpl
 
@@ -32,15 +30,4 @@ class MainViewModel : ViewModel() {
         value = prepareUser
     }
     val currentUser: LiveData<User> = _currentUser
-
-    //events
-
-    private val _prepareEventList = MutableLiveData<List<Event>>().apply {
-        repository.getEventsList(object : CallBack<List<Event>> {
-            override fun onResult(result: List<Event>) {
-                value = result
-            }
-        })
-    }
-    val prepareEventList: LiveData<List<Event>> = _prepareEventList
 }
