+++ app/src/main/java/ru/vdv/myapp/myreadersdiary/ui/readingProcess/ReadingProcessViewModel.kt
@@ -26,7 +26,7 @@ class ReadingProcessViewModel : ViewModel() {
 
     val text: LiveData<String> = _text
 
-    fun testGet() {
+    fun testGet(){
         Log.d("Моя проверка", "VM / Запущен метод TestGet")
         repository.getListOfBooks(object : CallBack<List<Book>> {
             override fun onResult(value: List<Book>) {
@@ -34,8 +34,7 @@ class ReadingProcessViewModel : ViewModel() {
             }
         })
     }
-
-    fun testPost() {
+    fun testPost(){
         Log.d("Моя проверка", "VM / Запущен метод TestPost")
         repository.postBook(object : CallBack<Any> {
             override fun onResult(result: Any) {
@@ -48,8 +47,7 @@ class ReadingProcessViewModel : ViewModel() {
             }
         })
     }
-
-    fun testPatch() {
+    fun testPatch(){
         Log.d("Моя проверка", "VM / Запущен метод TestPatch")
         repository.patchBook(object : CallBack<Any> {
             override fun onResult(result: Any) {
@@ -62,8 +60,7 @@ class ReadingProcessViewModel : ViewModel() {
             }
         })
     }
-
-    fun testDelete() {
+    fun testDelete(){
         Log.d("Моя проверка", "VM / Запущен метод TestDelete")
         repository.deleteBook(object : CallBack<Any> {
             override fun onResult(result: Any) {
@@ -76,8 +73,7 @@ class ReadingProcessViewModel : ViewModel() {
             }
         })
     }
-
-    fun testPut() {
+    fun testPut(){
         Log.d("Моя проверка", "VM / Запущен метод TestPut")
         repository.putBook(object : CallBack<Any> {
             override fun onResult(result: Any) {
@@ -90,8 +86,7 @@ class ReadingProcessViewModel : ViewModel() {
             }
         })
     }
-
-    fun testHead() {
+    fun testHead(){
         Log.d("Моя проверка", "VM / Запущен метод TestHead")
         repository.headBook(object : CallBack<Any> {
             override fun onResult(result: Any) {
@@ -104,8 +99,7 @@ class ReadingProcessViewModel : ViewModel() {
             }
         })
     }
-
-    fun testOptions() {
+    fun testOptions(){
         Log.d("Моя проверка", "VM / Запущен метод TestOptions")
         repository.optionsBook(object : CallBack<Any> {
             override fun onResult(result: Any) {
@@ -118,8 +112,7 @@ class ReadingProcessViewModel : ViewModel() {
             }
         })
     }
-
-    fun testHttp() {
+    fun testHttp(){
         Log.d("Моя проверка", "VM / Запущен метод TestHttp")
         repository.httpBook(object : CallBack<Any> {
             override fun onResult(result: Any) {
