--- app/src/main/res/layout/event_list_item.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="64dp"
-    android:layout_margin="4dp"
-    android:elevation="2dp"
-    android:padding="4dp"
-    app:cardCornerRadius="8dp"
-    app:cardElevation="2dp"
-    app:contentPadding="0dp">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ImageView
-            android:id="@+id/imageView_book_cover"
-            android:layout_width="50dp"
-            android:layout_height="64dp"
-            android:contentDescription="@string/ImageView_contentDescription_book_cover"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:srcCompat="@drawable/book_plug" />
-
-        <TextView
-            android:id="@+id/textView_event_title"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="4dp"
-            android:layout_marginEnd="4dp"
-            android:text="@string/textView_event_title_plug"
-            android:textAlignment="textStart"
-            android:textSize="16sp"
-            android:textStyle="bold|italic"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/imageView_book_cover"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/textView_fix_data_event"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="4dp"
-            android:text="@string/textView_fix_data_event_plug"
-            app:layout_constraintStart_toEndOf="@id/imageView_book_cover"
-            app:layout_constraintTop_toBottomOf="@id/textView_event_title" />
-
-        <TextView
-            android:id="@+id/textView_event_body"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="4dp"
-            android:text="@string/textView_event_body_plug"
-            android:textStyle="italic"
-            app:layout_constraintStart_toEndOf="@id/imageView_book_cover"
-            app:layout_constraintTop_toBottomOf="@id/textView_fix_data_event" />
-
-        <ImageView
-            android:id="@+id/imageView_event_add_book"
-            android:layout_width="32dp"
-            android:layout_height="32dp"
-            android:layout_margin="16dp"
-            android:contentDescription="@string/ImageView_contentDescription_sign_of_approval"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:srcCompat="@drawable/ic_baseline_library_add_24" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</com.google.android.material.card.MaterialCardView>
